seperation of opfunctions
